This optimization is pretty much basic.

Shebang-fixer code optimization:
Instead of iterating all possible files in massage directory we explicitly search for text files having `#` symbol at the start of any line.
This optimization reduces shebang-fixing from ~1m 33s to ~80ms while building `kf6-breeze-icons` on my machine.

OpenMP symbols scanning:
Instead of iterating all possible files we explicitly search files having elf magic at the start of any line.
This optimization reduces openmp-scanning from ~6s 700 ms to ~290ms while building `kf6-breeze-icons` on my machine.

The whole time of building `kf6-breeze-icons` was reduced from ~2m 10s to ~35s.

I am pretty much sure other packages will have similar results.

CC @truboxl since you wrote the elf-openmp-scanning code.