@TomJo2000 2025-04-27
> I think comments are unnecessary bloat when the code itself is self explainatory

That is absolutely not the mentality you should have towards comments.
- 1.) They are not executed, so they consume so runtime resources other than a couple bytes of storage space, which you can spare.
- 2.) Comments aren't for you (the author), they're for everyone who wants to read, understand, and maybe work of this code in the future.
And yes, that may even include future you.
But what'd obvious or "self explanatory" to the author (current you), may be a stumbling block or break in the flow for future you, or someone else reading this later.