@twaik 2025-04-29
Ok, now massaging step lasts slightly slower than with `perl` oneliners, but it is much more readable and and does not have external dependencies.

I did a little "benchmarking" with 
```
sudo sync; sudo sysctl -w vm.drop_caches=3; rm -rf ~/.termux-build/kf6-breeze-icons/host-build/; time ./build-package.sh -I -f kf6-breeze-icons -s
```
To build package with cached dependencies but with dropping disk page cache and measured `termux_step_massage`'s time.

Before:
```
real    1m41.697s
user    1m49.225s
sys     3m29.192s
```

After:
```
real    0m5.912s
user    0m10.221s
sys     0m1.329s
```

This time I hope it is a final revision of this PR.