@truboxl 2025-04-29
`$pattern` is used after line 73 which you have deleted and replaced with a duplicated function. Might as well delete the whole line 63 to line 73.