@twaik 2025-04-29
Ooooh, ok. Now I get it. I did not remove it because I pasted agnostic-apollo's code without reviewing other parts of code it could duplicate.